+++ src/classes/IDPackCell.cpp
@@ -42,7 +42,7 @@ bool IDPackCell::init(const std::string& name, double points, const std::vector<
     auto viewMenu = CCMenu::create();
     auto viewButton = CCMenuItemExt::createSpriteExtra(viewSprite, [this, levels](auto) {
         CCDirector::get()->pushScene(CCTransitionFade::create(0.5f, LevelBrowserLayer::scene(GJSearchObject::create(SearchType::Type19,
-            ranges::reduce<std::string>(levels, [](std::string& str, int level) { str += (str.empty() ? "" : ",") + std::to_string(level); })))));
+            ranges::reduce<std::string>(levels, [](std::string& str, int level) { str += (str.empty() ? "" : ",") + fmt::to_string(level); })))));
     });
     viewButton->setID("view-button");
     viewMenu->addChild(viewButton);